+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changes
 - Revoked previous backend acceptance to display privacy policy
+- Added (ᵔᴥᵔ) as a trigger for the :dog: emoji
 
 ### Fixes